Propagate asset information to every milestone and commission year
Adding an asset with a keyword from a milestone or commission year will
automatically propagate that information to every revelant year. This
allows setting asset properties without having to now details about
milestone and commission schema.

It is now also necessary to inform which years are milestone years.
Not attaching a profile to an asset in a given year will mark that
year as milestone.